+++ docs/CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [Beta Release]
+
+## [Beta Release] - [1.5.1] - 20200918 0400 UTC - Beta
+ - Fixed Admin Index plugin population
+ - Fixed Admin Menu plugin population
+ - added missing variable to admin->users
+ - repackaged 1.5.1 for beta
+
+## [Beta Release] - [1.5.1] - 20200918 0230 UTC - Beta
  - Release 1.5.1 Beta for testing
 
 ## [Unreleased] - [1.5.1] - 20200918 0130 UTC - Development